% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eim-package.R
\name{as.matrix.eim}
\alias{as.matrix.eim}
\title{Returns the eim's model probability matrix.}
\usage{
\method{as.matrix}{eim}(object)
}
\arguments{
\item{object}{}
}
\value{
(matrix) A matrix containing the estimated probabilities.
}
\description{
Extracts the probability matrix from the model, making it able to manipulate it as
a matrix
}
\examples{
model <- eim$new(X = matrix(1:15, 5, 3), W = matrix(1:50, 5, 10))
predict(model)
ncol(as.matrix(model)) # Returns the column of probabilities

}
