% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instancer.R
\name{simulate_votation}
\alias{simulate_votation}
\title{Simulates a votation, creating the candidates and group matrices and its results.}
\usage{
simulate_votation(
  num_ballots,
  num_candidates,
  num_demographics,
  ballot_voters,
  lambda = 0.5
)
}
\arguments{
\item{num_ballots}{(integer(1)) The number of ballot boxes \emph{("b")}.}

\item{num_candidates}{(integer(1)) The number of candidates \emph{("c")}.}

\item{num_demographics}{(integer(1)) The number of demographic groups \emph{("g")}.}

\item{ballot_voters}{(integer(num_ballots)) A vector of size \code{num_ballots} with the amount of votes per ballot box.}

\item{lambda}{(numeric(1)) Value between 0 and 1 that represents the group heterogeneity. Values near 0 are more heterogeneous,
but less realistic. The same goes the other way around. (default: 0.5)}
}
\value{
A list with components:
\item{X}{A matrix (b x c) with candidate votes per ballot box.}
\item{W}{A matrix (b x g) with demographic votes per ballot box.}
\item{ground_truth}{A matrix (g x c) with the estimated (unseen) probabilities that a demographic group votes for a given candidate.}
}
\description{
Given the number of ballots, groups, candidates and votes per ballot it will simulate a votation.
Additionally, it generates the resulting matrix (of dimension g x c) that represents the unseen probabilities
that a demographic group votes for a given candidate. These probabilities are drawn from a Dirichlet distribution
and a \code{lambda} value which represents the heterogeneity of the groups.
}
\note{
A reminder from the Dirichlet distribution is that, an alpha = 1 represents a distribution that is
uniform around the mean. On the other side, alpha < 1 tends to sparse the probability vector, with many
values close to zero and one or few larger values. The inverse case happens when alpha > 1, tightening
the concentration among the mean.
}
\examples{
result <- simulate_votation(num_ballots = 10, num_candidates = 5, num_demographics = 3, ballot_voters = c(100, 10))
result$X # Candidate matrix (b x c)
result$W # Group matrix (b x g)
result$ground_truth # Probability matrix (g x c)

}
\seealso{
[random_samples()]
}
