% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eim-package.R
\name{sum.eim}
\alias{sum.eim}
\title{Returns the total amount of voters in the system.}
\usage{
\method{sum}{eim}(object)
}
\arguments{
\item{object}{An eim object.}
}
\value{
The amount of voters.
}
\description{
Given a initialized object, it returns the total amount of voters. It's equivalent
of running sum(object$W) or sum(object$X)
}
\examples{
model <- eim$new(X = matrix(1:15, 5, 3), W = matrix(1:50, 5, 10))
sum(model) # 1395

}
