% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eim-package.R
\name{summary.eim}
\alias{summary.eim}
\title{Summarize the object main attributes.}
\usage{
\method{summary}{eim}(object)
}
\arguments{
\item{A}{eim object}
}
\value{
(list) A list with the principal attributes
}
\description{
This method is an S3 styled wrapper around summary(), allowing to get, as a list,
		the main details of the object. Additional parameters are shown depending on the state of
		the object (computed/uncomputed).
}
\details{
Return a list with the principal attributes
}
\examples{
model <- eim$new(X = matrix(1:15, 5, 3), W = matrix(1:10, 2, 5))
a_list <- summary(model)
a_list$method # Not computed yet
a_list$groups # 2
a_list$ballots # 5
names(a_list) # "candidates" "groups" "ballots" "method" "probabilities" "logLikelihood"

}
