% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eim-package.R
\name{update.eim}
\alias{update.eim}
\title{Update an existing eim model with a new EM algorithm computation}
\usage{
\method{update}{eim}(object, ...)
}
\arguments{
\item{object}{An eim object.}

\item{...}{New parameters to pass to compute().}
}
\value{
Changes made on the updated object.
}
\description{
This function updates an object with a new Expected Maximization computation
		with other parameters.
}
\examples{
model <- eim$new(X = matrix(1:15, 5, 3), W = matrix(1:50, 5, 10))
predict(model)
update(model, "MVN PDF")
model$method # Returns 'MVN PDF'

}
